Add opt-in features to turn off/disable language features.

+++ tests/binary_ops.rs
@@ -1,15 +1,15 @@
-use rhai::{Engine, EvalAltResult};
+use rhai::{Engine, EvalAltResult, INT};
 
 #[test]
 fn test_binary_ops() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
-    assert_eq!(engine.eval::<i64>("10 % 4")?, 2);
-    assert_eq!(engine.eval::<i64>("10 << 4")?, 160);
-    assert_eq!(engine.eval::<i64>("10 >> 4")?, 0);
-    assert_eq!(engine.eval::<i64>("10 & 4")?, 0);
-    assert_eq!(engine.eval::<i64>("10 | 4")?, 14);
-    assert_eq!(engine.eval::<i64>("10 ^ 4")?, 14);
+    assert_eq!(engine.eval::<INT>("10 % 4")?, 2);
+    assert_eq!(engine.eval::<INT>("10 << 4")?, 160);
+    assert_eq!(engine.eval::<INT>("10 >> 4")?, 0);
+    assert_eq!(engine.eval::<INT>("10 & 4")?, 0);
+    assert_eq!(engine.eval::<INT>("10 | 4")?, 14);
+    assert_eq!(engine.eval::<INT>("10 ^ 4")?, 14);
 
     assert_eq!(engine.eval::<bool>("42 == 42")?, true);
     assert_eq!(engine.eval::<bool>("42 > 42")?, false);
