+++ tests/decrement.rs
@@ -1,10 +1,10 @@
-use rhai::{Engine, EvalAltResult};
+use rhai::{Engine, EvalAltResult, INT};
 
 #[test]
 fn test_decrement() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
-    assert_eq!(engine.eval::<i64>("let x = 10; x -= 7; x")?, 3);
+    assert_eq!(engine.eval::<INT>("let x = 10; x -= 7; x")?, 3);
 
     let r = engine.eval::<String>("let s = \"test\"; s -= \"ing\"; s");
